docs/source/requirements.txt
sphinx_rtd_theme
sphinx
sphinx-autobuild
requests
myst-parser
PyYAML
